Add files via upload
 BRIEF DESCRIPTION 
 
The mini-project entitled “PHARMACY MANAGEMENT SYSTEM “ is developed as a part of the fifth semester DBMS laboratory, for the partial fulfillment of the requirement for the BE( Information Science) course. 
 
 
Project Description 
The project entitled Pharmacy Management System this is developed for established Medical  store in the city. To manage all operations of the medical store this project is being developed. It  will have the entire basic module to manage the medical store operations. 
 
Objectives 
The main objective of the application is to automate the existing system of manually maintained records of the counter sales, purchases, reorder levels, Supplier and Customer monetary positions and other related transactions made by the seller. 

SCOPE 
This application can be used by any other store to automate the process of manually maintaining the records related to the subject of maintaining the stock and liquid flows.